+++ .gitignore
@@ -1,4 +1,4 @@
-/vendor/
+**/vendor/
 .pulumi
 **/bin/
 **/obj/
@@ -11,3 +11,4 @@ sdk/nodejs/node_modules
 *.iml
 
 yarn.lock
+**/pulumiManifest.go
