+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD (Unreleased)
-_(none)_
+* Upgrade to Pulumi v1.13.1
+* Upgrade to pulumi-terraform-bridge v1.8.4
+* Refactor layout to support Go modules
 
 ---
