+++ provider/cmd/pulumi-resource-postgresql/.gitignore
@@ -1,0 +1,2 @@
+schema.go
+schema.json
