+++ provider/cmd/pulumi-resource-postgresql/Pulumi.yaml
