+++ provider/cmd/pulumi-resource-postgresql/generate.go
