+++ provider/cmd/pulumi-resource-postgresql/main.go
@@ -12,15 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//go:generate go run ./generate.go
+
 package main
 
 import (
-	"github.com/pulumi/pulumi-terraform-bridge/pkg/tfgen"
+	"github.com/pulumi/pulumi-terraform-bridge/pkg/tfbridge"
 
-	postgresql "github.com/pulumi/pulumi-postgresql"
-	"github.com/pulumi/pulumi-postgresql/pkg/version"
+	postgresql "github.com/pulumi/pulumi-postgresql/provider"
+	"github.com/pulumi/pulumi-postgresql/provider/pkg/version"
 )
 
 func main() {
-	tfgen.Main("postgresql", version.Version, postgresql.Provider())
+	// Modify the path to point to the new provider
+	tfbridge.Main("postgresql", version.Version, postgresql.Provider(), pulumiSchema)
 }
