+++ provider/cmd/pulumi-tfgen-postgresql/main.go
@@ -12,18 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:generate go run ./generate.go
-
 package main
 
 import (
-	"github.com/pulumi/pulumi-terraform-bridge/pkg/tfbridge"
+	"github.com/pulumi/pulumi-terraform-bridge/pkg/tfgen"
 
-	postgresql "github.com/pulumi/pulumi-postgresql"
-	"github.com/pulumi/pulumi-postgresql/pkg/version"
+	postgresql "github.com/pulumi/pulumi-postgresql/provider"
+	"github.com/pulumi/pulumi-postgresql/provider/pkg/version"
 )
 
 func main() {
-	// Modify the path to point to the new provider
-	tfbridge.Main("postgresql", version.Version, postgresql.Provider(), pulumiSchema)
+	tfgen.Main("postgresql", version.Version, postgresql.Provider())
 }
