+++ scripts/get-py-version
@@ -4,4 +4,4 @@ set -o nounset -o errexit -o pipefail
 SCRIPT_ROOT="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 NPM_VERSION="$($SCRIPT_ROOT/get-version "$@")"
 
-go run "${SCRIPT_ROOT}/get-py-version.go" "${NPM_VERSION}"
+GO111MODULE=on go run "${SCRIPT_ROOT}/get-py-version.go" "${NPM_VERSION}"
