+++ scripts/go.mod
@@ -1,0 +1,5 @@
+module github.com/pulumi/pulumi-postgresql/scripts
+
+go 1.13
+
+require github.com/pulumi/pulumi/sdk v1.14.0
