+++ sdk/go/postgresql/grant.go
@@ -133,4 +133,3 @@ type GrantArgs struct {
 func (GrantArgs) ElementType() reflect.Type {
 	return reflect.TypeOf((*grantArgs)(nil)).Elem()
 }
-
